requirements.txt
alembic==1.13.1
cryptography==42.0.8
fastapi==0.111.0
httptools==0.6.1
pip-chill==1.0.3
pyjwt==2.8.0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.32.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
